pushapp/requirement.txt
asgiref==3.2.10
attrs==20.2.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
Django==3.1.2
djangorestframework==3.12.1
fcm-django==0.3.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
numpy==1.19.2
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.7
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.24.0
scipy==1.5.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
treq==20.9.0
Twisted==20.3.0
txFCM==0.0.1
urllib3==1.25.10
zope.interface==5.1.2

